refactor pages with main layout component
