+++ src/parts/Logo.js
